+++ Cargo.toml
@@ -4,7 +4,7 @@ version = "0.0.2"
 edition = "2021"
 license = "MIT"
 description = "Lerni content framework"
-repository = "https://github.com/lerni-edu/lerni"
+repository = "https://github.com/noviga/lerni"
 include = ["src/**/*.rs", "examples/**/*.rs"]
 
 [dependencies]
@@ -15,7 +15,6 @@ gloo-utils = "0.2.0"
 leptos = { version = "0.5.1", features = ["csr"] }
 leptos-use = "0.8.0"
 rand = { version = "0.8.5", default-features = false, features = ["getrandom"] }
-yew = { version = "0.21.0", features = ["csr"] }
 wasm-bindgen = "0.2.84"
 
 [dependencies.web-sys]
@@ -41,38 +40,6 @@ crate-type = ["cdylib"]
 
 [[example]]
 name = "hello_world"
-crate-type = ["cdylib"]
-
-[[example]]
-name = "ng_blur"
-crate-type = ["cdylib"]
-
-[[example]]
-name = "ng_buttons"
-crate-type = ["cdylib"]
-
-[[example]]
-name = "ng_grid"
-crate-type = ["cdylib"]
-
-[[example]]
-name = "ng_hello_world"
-crate-type = ["cdylib"]
-
-[[example]]
-name = "ng_pointer"
-crate-type = ["cdylib"]
-
-[[example]]
-name = "ng_rows_cols"
-crate-type = ["cdylib"]
-
-[[example]]
-name = "ng_svg"
-crate-type = ["cdylib"]
-
-[[example]]
-name = "ng_text"
 crate-type = ["cdylib"]
 
 [[example]]
