+++ examples/blur.rs
@@ -1,23 +1,21 @@
-use lerni::{properties::Color, widgets::*};
-use wasm_bindgen::prelude::wasm_bindgen;
-use yew::prelude::*;
+use leptos::*;
+use lerni::*;
 
-#[function_component]
-pub fn Blur() -> Html {
-    let blur = use_state(|| false);
-    let onclick = {
-        let blur = blur.clone();
-        Callback::from(move |_| blur.set(!*blur))
-    };
+#[component]
+pub fn Blur() -> impl IntoView {
+    let (blur, set_blur) = create_signal(false);
+    let on_click = move |_| set_blur.set(!blur.get());
 
-    html! {
-        <Slide blur={ *blur } background_color={ Color::MistyRose }>
-            <Button text="Blur ON/OFF" { onclick } />
+    view! {
+        <Slide blur=blur.into() background_color=Color::MistyRose>
+            <Button on_click=on_click>
+                "Blur " {move || if blur.get() { "ON" } else { "OFF" }}
+            </Button>
         </Slide>
     }
 }
 
 #[wasm_bindgen(start)]
 pub fn main() {
-    lerni::start::<Blur>();
+    lerni::start(Blur);
 }
