+++ examples/hello_world.rs
@@ -1,18 +1,21 @@
-use lerni::{properties::Color, widgets::*};
-use wasm_bindgen::prelude::wasm_bindgen;
-use yew::prelude::*;
+use leptos::*;
+use lerni::*;
 
-#[function_component]
-pub fn HelloWorld() -> Html {
-    html! {
+#[component]
+pub fn HelloWorld() -> impl IntoView {
+    view! {
         <SlideShow>
-            <Slide background_color={ Color::MistyRose }><Label text="Hello →"  /></Slide>
-            <Slide background_color={ Color::PaleGreen }><Label text="← World!" /></Slide>
+            <Slide background_color=Color::MistyRose>
+                <Label>"Hello →"</Label>
+            </Slide>
+            <Slide background_color=Color::PaleGreen>
+                <Label>"← World!"</Label>
+            </Slide>
         </SlideShow>
     }
 }
 
 #[wasm_bindgen(start)]
 pub fn main() {
-    lerni::start::<HelloWorld>();
+    lerni::start(HelloWorld);
 }
