--- examples/ng_hello_world.rs
@@ -1,21 +1,0 @@
-use leptos::*;
-use lerni::ng::*;
-
-#[component]
-pub fn HelloWorld() -> impl IntoView {
-    view! {
-        <SlideShow>
-            <Slide background_color=Color::MistyRose>
-                <Label>"Hello →"</Label>
-            </Slide>
-            <Slide background_color=Color::PaleGreen>
-                <Label>"← World!"</Label>
-            </Slide>
-        </SlideShow>
-    }
-}
-
-#[wasm_bindgen(start)]
-pub fn main() {
-    lerni::ng::start(HelloWorld);
-}
