+++ examples/rows_cols.rs
@@ -1,25 +1,21 @@
-use lerni::{
-    properties::{Align, VAlign},
-    widgets::*,
-};
-use wasm_bindgen::prelude::wasm_bindgen;
-use yew::prelude::*;
+use leptos::*;
+use lerni::*;
 
-#[function_component]
-pub fn RowsCols() -> Html {
-    html! {
+#[component]
+pub fn RowsCols() -> impl IntoView {
+    view! {
         <Slide>
-            <Row border_width=4 stretch={ vec![1, 1, 4, 1, 1] } padding=20>
-                <Label text="1" />
-                <Button text="2" align={ Align::Fill } valign={ VAlign::Fill } />
-                <Label text="3" bold=true />
-                <Column border_width=4 stretch={ vec![1, 2, 3, 4] } spacing=20>
-                    <Label text="4" />
-                    <Label text="5" />
-                    <Button text="6" align={ Align::Fill } valign={ VAlign::Fill } />
-                    <Label text="7" />
+            <Row border_width=4 stretch=vec![1, 1, 4, 1, 1] padding=20>
+                <Label>"1"</Label>
+                <Button align=Align::Fill valign=VAlign::Fill on_click=|_| ()>"2"</Button>
+                <Label bold=true>"3"</Label>
+                <Column border_width=4 stretch=vec![1, 2, 3, 4] spacing=20>
+                    <Label>"4"</Label>
+                    <Label>"5"</Label>
+                    <Button align=Align::Fill valign=VAlign::Fill on_click=|_| ()>"6"</Button>
+                    <Label>"7"</Label>
                 </Column>
-                <Label text="8" />
+                <Label>"8"</Label>
             </Row>
         </Slide>
     }
@@ -27,5 +23,5 @@ pub fn RowsCols() -> Html {
 
 #[wasm_bindgen(start)]
 pub fn main() {
-    lerni::start::<RowsCols>();
+    lerni::start(RowsCols);
 }
