+++ examples/svg.rs
@@ -1,24 +1,24 @@
-use lerni::{
-    properties::{Align, VAlign},
-    widgets::*,
-};
-use wasm_bindgen::prelude::wasm_bindgen;
-use yew::prelude::*;
+use leptos::*;
+use lerni::*;
 
-#[function_component]
-pub fn SvgExample() -> Html {
-    html! {
-        <Slide background_image="/img/lerni-bg.svg">
+#[component]
+pub fn SvgExample() -> impl IntoView {
+    view! {
+        <Slide background_image="/img/lerni-bg.svg".into()>
             <Grid cols=3 rows=3 border_width=4 padding=20>
-                <Svg width=128 height=64>{ include!("logo.svg-rs") }</Svg>
-                <Svg width=128 height=64 scale=3.0>{ include!("logo.svg-rs") }</Svg>
-                <Svg width=128 height=64 scale=2.0 align={ Align::Left } valign={ VAlign::Top }>{ include!("logo.svg-rs") }</Svg>
-                <Svg width=128 height=64 align={ Align::Fill }>{ include!("logo.svg-rs") }</Svg>
-                <Svg width=128 height=64 flip_x=true>{ include!("logo.svg-rs") }</Svg>
-                <Svg width=128 height=64 scale=2.0>{ include!("logo.svg-rs") }</Svg>
-                <Svg width=128 height=64 valign={ VAlign::Bottom }>{ include!("logo.svg-rs") }</Svg>
-                <Svg width=128 height=64 flip_y=true>{ include!("logo.svg-rs") }</Svg>
-                <Svg width=128 height=64 flip_x=true flip_y=true scale=2.0 align={ Align::Right } valign={ VAlign::Bottom }>{ include!("logo.svg-rs") }</Svg>
+                <SvgFile width=128 height=64 src=include_str!("logo.svg") />
+                <Svg width=128 height=64 scale=3.0>{include!("logo.svg-rs")}</Svg>
+                <Svg width=128 height=64 scale=2.0 align=Align::Left valign=VAlign::Top>
+                    {include!("logo.svg-rs")}
+                </Svg>
+                <SvgFile width=128 height=64 align=Align::Fill src=include_str!("logo.svg")/>
+                <Svg width=128 height=64 flip_x=true>{include!("logo.svg-rs")}</Svg>
+                <Svg width=128 height=64 scale=2.0>{include!("logo.svg-rs")}</Svg>
+                <Svg width=128 height=64 valign=VAlign::Bottom>{ include!("logo.svg-rs")}</Svg>
+                <Svg width=128 height=64 flip_y=true>{include!("logo.svg-rs")}</Svg>
+                <Svg width=128 height=64 flip_x=true flip_y=true scale=2.0 align=Align::Right valign=VAlign::Bottom>
+                    {include!("logo.svg-rs")}
+                </Svg>
             </Grid>
         </Slide>
     }
@@ -26,5 +26,5 @@ pub fn SvgExample() -> Html {
 
 #[wasm_bindgen(start)]
 pub fn main() {
-    lerni::start::<SvgExample>();
+    lerni::start(SvgExample);
 }
