--- src/ng.rs
@@ -1,75 +1,0 @@
-//! New generation Lerni based on Leptos.
-
-pub use wasm_bindgen::prelude::wasm_bindgen;
-
-mod align;
-pub use align::{Align, VAlign};
-
-mod button;
-pub use button::Button;
-
-mod color;
-pub use color::Color;
-
-mod column;
-pub use column::Column;
-
-mod frame;
-pub use frame::*;
-
-mod grid;
-pub use grid::Grid;
-
-mod label;
-pub use label::Label;
-
-mod row;
-pub use row::Row;
-
-mod slide;
-pub use slide::Slide;
-
-mod slideshow;
-pub use slideshow::SlideShow;
-
-mod svg;
-pub use svg::{Svg, SvgFile};
-
-mod text;
-pub use text::Text;
-
-use leptos::*;
-
-/// Additional information provided to all slides.
-#[derive(Clone, Copy, Default, Debug, PartialEq)]
-pub struct Metadata {
-    /// Visibility flag.
-    pub visible: bool,
-    /// Teacher mode flag.
-    pub teacher_mode: bool,
-    /// Pointer on/off flag.
-    pub pointer: bool,
-}
-
-/// Calculates the width of the slide.
-pub fn calc_width(margin: i32) -> i32 {
-    let elem = web_sys::window()
-        .and_then(|win| win.document())
-        .and_then(|doc| doc.document_element());
-    if let Some(elem) = elem {
-        let width = elem.client_width();
-        let height = elem.client_height();
-        width.min((height - margin) * 16 / 9)
-    } else {
-        0
-    }
-}
-
-/// The main entry point.
-pub fn start<F, N>(f: F)
-where
-    F: Fn() -> N + 'static,
-    N: IntoView,
-{
-    leptos::mount_to_body(f);
-}
