--- src/ng/color.rs
@@ -1,321 +1,0 @@
-use derive_more::Display;
-use leptos::{Attribute, IntoAttribute};
-
-// TODO: Define all colors
-
-/// Helper type to work with colors.
-#[derive(Clone, Copy, Default, Display, PartialEq)]
-pub enum Color {
-    /// No color.
-    #[default]
-    None,
-    /// Arbitrary color with the string value.
-    #[display(fmt = "{_0}")]
-    Value(&'static str),
-
-    /// Alice Blue color.
-    AliceBlue,
-    /// Antique White color.
-    AntiqueWhite,
-    ///
-    Aqua,
-    ///
-    Aquamarine,
-    ///
-    Azure,
-    ///
-    Aeige,
-    ///
-    Bisque,
-    ///
-    Black,
-    ///
-    BlanchedAlmond,
-    ///
-    Blue,
-    ///
-    BlueViolet,
-    ///
-    Brown,
-    ///
-    BurlyWood,
-    ///
-    CadetBlue,
-    ///
-    Chartreuse,
-    ///
-    Chocolate,
-    ///
-    Coral,
-    ///
-    CornflowerBlue,
-    ///
-    Cornsilk,
-    ///
-    Crimson,
-    ///
-    Cyan,
-    ///
-    DarkBlue,
-    ///
-    DarkCyan,
-    ///
-    DarkGoldenrod,
-    ///
-    DarkGray,
-    ///
-    DarkGreen,
-    ///
-    DarkGrey,
-    ///
-    DarkKhaki,
-    ///
-    DarkMagenta,
-    ///
-    DarkOliveGreen,
-    ///
-    DarkOrange,
-    ///
-    DarkOrchid,
-    ///
-    DarkRed,
-    ///
-    DarkSalmon,
-    ///
-    DarkSeaGreen,
-    ///
-    DarkSlateBlue,
-    ///
-    DarkSlateGray,
-    ///
-    DarkSlateGrey,
-    ///
-    ///
-    DarkTurquoise,
-    ///
-    DarkViolet,
-    ///
-    DeepPink,
-    ///
-    DeepSkyBlue,
-    ///
-    DimGray,
-    ///
-    DimGrey,
-    ///
-    DodgerBlue,
-    ///
-    Firebrick,
-    ///
-    FloralWhite,
-    ///
-    ForestGreen,
-    ///
-    Fuchsia,
-    ///
-    Gainsboro,
-    ///
-    GhostWhite,
-    ///
-    Gold,
-    ///
-    Goldenrod,
-    ///
-    Gray,
-    ///
-    Green,
-    ///
-    GreenYellow,
-    ///
-    Grey,
-    ///
-    Honeydew,
-    ///
-    HotPink,
-    ///
-    IndianRed,
-    ///
-    Indigo,
-    ///
-    Ivory,
-    ///
-    Khaki,
-    ///
-    Lavender,
-    ///
-    LavenderBlush,
-    ///
-    LawnGreen,
-    ///
-    Lemonchiffon,
-    ///
-    LightBlue,
-    ///
-    LightCoral,
-    ///
-    LightCyan,
-    ///
-    LightGoldenrodYellow,
-    ///
-    LightGray,
-    ///
-    LightGreen,
-    ///
-    LightGrey,
-    ///
-    LightPink,
-    ///
-    LightSalmon,
-    ///
-    LightSeaGreen,
-    ///
-    LightSkyBlue,
-    ///
-    LightSlateGray,
-    ///
-    LightSlateGrey,
-    ///
-    LightsteelBlue,
-    ///
-    LightYellow,
-    ///
-    Lime,
-    ///
-    LimeGreen,
-    ///
-    Linen,
-    ///
-    Magenta,
-    ///
-    Maroon,
-    ///
-    MediumAquamarine,
-    ///
-    MediumBlue,
-    ///
-    MediumOrchid,
-    ///
-    MediumPurple,
-    ///
-    MediumSeaGreen,
-    ///
-    MediumSlateBlue,
-    ///
-    MediumSpringGreen,
-    ///
-    MediumTurquoise,
-    ///
-    MediumVioletRed,
-    ///
-    MidnightBlue,
-    ///
-    MintCream,
-    ///
-    MistyRose,
-    ///
-    Moccasin,
-    ///
-    NavajoWhite,
-    ///
-    Navy,
-    ///
-    Oldlace,
-    ///
-    Olive,
-    ///
-    Olivedrab,
-    ///
-    Orange,
-    ///
-    OrangeRed,
-    ///
-    Orchid,
-    ///
-    PaleGoldenrod,
-    ///
-    PaleGreen,
-    ///
-    PaleTurquoise,
-    ///
-    PaleVioletRed,
-    ///
-    Papayawhip,
-    ///
-    Peachpuff,
-    ///
-    Peru,
-    ///
-    Pink,
-    ///
-    Plum,
-    ///
-    PowderBlue,
-    ///
-    Purple,
-    ///
-    Red,
-    ///
-    RosyBrown,
-    ///
-    RoyalBlue,
-    ///
-    SaddleBrown,
-    ///
-    Salmon,
-    ///
-    SandyBrown,
-    ///
-    SeaGreen,
-    ///
-    Seashell,
-    ///
-    Sienna,
-    ///
-    Silver,
-    ///
-    SkyBlue,
-    ///
-    SlateBlue,
-    ///
-    SlateGray,
-    ///
-    SlateGrey,
-    ///
-    Snow,
-    ///
-    SpringGreen,
-    ///
-    SteelBlue,
-    ///
-    Tan,
-    ///
-    Teal,
-    ///
-    Thistle,
-    ///
-    Tomato,
-    ///
-    Turquoise,
-    ///
-    Violet,
-    ///
-    Wheat,
-    ///
-    White,
-    ///
-    WhiteSmoke,
-    ///
-    Yellow,
-    /// Yellow Green color
-    YellowGreen,
-}
-
-impl IntoAttribute for Color {
-    fn into_attribute(self) -> Attribute {
-        Attribute::String(self.to_string().into())
-    }
-
-    fn into_attribute_boxed(self: Box<Self>) -> leptos::Attribute {
-        self.into_attribute()
-    }
-}
